Preparing for 1.0
Fixed issue with Pyinstaller binaries where the scripts window would not work properly due to not being able to find the scripts folder as a side effect of the way Pyinstaller packages programs.
Plugin functionality has been removed due to it being too much work for a first release, it would've required a bunch of work to be done to ensure the security of the program.
Finally, an update system has been made where after each launch, the version.txt file is compared to the one found in the Github master branch, if a difference is found a new "update released" window that prompts the user to click yes to go to the latest release or no to stay on the current version they have.